scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.1
Pillow==10.3.0
ImageHash==4.3.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
tqdm==4.66.4
